[Evals] Refactor tasks, add linting and CI  (#47)
Refactors tasks, adds linters and CI with Github workflows. 

## Task refactoring
With the Evaluation suite adding more and more tasks and supporting more models, there's a need for modularization. This PR is focused on refactoring task implementation and decoupling task configuration and implementation logic. 

## Linting
I've added basic linting support with ruff and black with pre-commit hooks for ease of use. 

## CI 
I've made a simple `tests` folder and added CI jobs - one for code quality and one for tests-  that run per commit. 

## Misc
Upgraded to vllm 0.7.0 - this was needed because there were fixes for `tokenizer_revision` args in 0.6.6 (or so) that we use for non-reasoning tasks.