[evals] Add support for scaling evals and inference with ray (#63)
This PR adds support for using ray to speed up evals and data generation. Currently we use are using a preliminary version of using ray data + vllm while we wait for the code at `ray.llm` to be fully open sourced (coming in the next 1-2 weeks), after which we will migrate over.